refactor(cachekey): move internal/cachekey to core/cachekey
Users need to be able to import the cache key helpers, so the package
can no longer live under the repo-private internal/ tree.
