+++ controller/gettargetgraph.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/uber/tango/core/common"
 	"github.com/uber/tango/entity"
-	"github.com/uber/tango/internal/cachekey"
+	"github.com/uber/tango/core/cachekey"
 	"github.com/uber/tango/internal/mapper"
 
 	"github.com/uber/tango/core/storage"
